Thumbnail modified

+++ src/main/java/com/cobong/yuja/service/board/BoardServiceImpl.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -66,22 +67,25 @@ public class BoardServiceImpl implements BoardService {
 		} 
 		
 		if(dto.getThumbnailId() != null && dto.getThumbnailId() != 0L) {
-			Thumbnail thumbnail = thumbnailRepository.findById(dto.getThumbnailId()).orElseThrow(() -> new IllegalAccessError("썸네일이 존재하지 않음 " + dto.getThumbnailId()));
-			if(!thumbnail.isFlag()) {
-				File temp = new File(thumbnail.getTempPath());
-				File dest = new File(thumbnail.getUploadPath());
-				File origTemp = new File(thumbnail.getOriginalFileTemp());
-				File origDest = new File(thumbnail.getOriginalFileDest());
-				try {
-					Files.move(temp, dest);
-					Files.move(origTemp, origDest);
-				} catch (IOException e) {
-					e.printStackTrace();
+			Optional<Thumbnail> thumbnailtodel = thumbnailRepository.findById(dto.getThumbnailId());
+			if(thumbnailtodel.isPresent()) {
+				Thumbnail thumbnail = thumbnailtodel.get();
+				if(!thumbnail.isFlag()) {
+					File temp = new File(thumbnail.getTempPath());
+					File dest = new File(thumbnail.getUploadPath());
+					File origTemp = new File(thumbnail.getOriginalFileTemp());
+					File origDest = new File(thumbnail.getOriginalFileDest());
+					try {
+						Files.move(temp, dest);
+						Files.move(origTemp, origDest);
+					} catch (IOException e) {
+						e.printStackTrace();
+					}
+					thumbnail.completelySave();
 				}
-				thumbnail.completelySave();
+				thumbnail.addBoard(board2);
+				thumbnailRepository.save(thumbnail);
 			}
-			thumbnail.addBoard(board2);
-			thumbnailRepository.save(thumbnail);
 		}
 		return new BoardResponseDto().entityToDto(board2);
 	}
@@ -102,10 +106,14 @@ public class BoardServiceImpl implements BoardService {
 		for(BoardAttach boardAttach: attaches) {
 			boardAttachesToSend.add(boardAttach.getFileName());
 		}
-		Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(bno);
+		
+		Optional<Thumbnail> thumbnailtodel = thumbnailRepository.findByBoardBoardId(bno);
 		String thumbnailOrig = "";
-		if(thumbnail.getOriginalFileDest() != null && thumbnail.getOriginalFileDest().length() != 0) {
-			thumbnailOrig += thumbnail.getOriginalFileDest();
+		if(thumbnailtodel.isPresent()) {
+			Thumbnail thumbnail = thumbnailtodel.get();
+			if(thumbnail.getOriginalFileDest() != null && thumbnail.getOriginalFileDest().length() != 0) {
+				thumbnailOrig += thumbnail.getOriginalFileDest();
+			}
 		}
 		/***
 		 * findById일때는 썸네일 테이블에서 오리지널 파일을 보내준다
@@ -143,23 +151,27 @@ public class BoardServiceImpl implements BoardService {
 			}
 		}
 		
-		Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(bno);
-		File thumbToDel = new File(thumbnail.getUploadPath());
-		if(thumbToDel.exists()) {
-			thumbToDel.delete();
-		} else {
-			/***
-			 * 파일 존재하지 않을시 예외처리 필요
-			 */
-		}
 		
-		File origToDel = new File(thumbnail.getOriginalFileDest());
-		if(origToDel.exists()) {
-			origToDel.delete();
-		} else {
-			/***
-			 * 파일 존재하지 않을시 예외처리 필요
-			 */
+		Optional<Thumbnail> thumbnailtodel = thumbnailRepository.findByBoardBoardId(bno);
+		if(thumbnailtodel.isPresent()) {
+			Thumbnail thumbnail = thumbnailtodel.get();
+			File thumbToDel = new File(thumbnail.getUploadPath());
+			if(thumbToDel.exists()) {
+				thumbToDel.delete();
+			} else {
+				/***
+				 * 파일 존재하지 않을시 예외처리 필요
+				 */
+			}
+			
+			File origToDel = new File(thumbnail.getOriginalFileDest());
+			if(origToDel.exists()) {
+				origToDel.delete();
+			} else {
+				/***
+				 * 파일 존재하지 않을시 예외처리 필요
+				 */
+			}			
 		}
 		
 		boardRepository.deleteById(bno);
@@ -207,7 +219,7 @@ public class BoardServiceImpl implements BoardService {
 			attachRepository.save(boardAttach);
 		}
 		
-		Thumbnail origThumbnail = thumbnailRepository.findByBoardBoardId(bno);
+		Thumbnail origThumbnail = thumbnailRepository.findByBoardBoardId(bno).orElseThrow(() -> new IllegalAccessError("썸네일이 존재하지 않음 " + bno));
 		if(origThumbnail != null) {
 			if(boardUpdateRequestDto.getThumbnailId() != origThumbnail.getThumbnailId()) {
 				File thumbToDel = new File(origThumbnail.getUploadPath());
@@ -228,22 +240,25 @@ public class BoardServiceImpl implements BoardService {
 				}
 				
 				if(boardUpdateRequestDto.getThumbnailId() != null && boardUpdateRequestDto.getThumbnailId() != 0L) {
-					Thumbnail newThumb = thumbnailRepository.findById(boardUpdateRequestDto.getThumbnailId()).orElseThrow(() -> new IllegalAccessError("썸네일이 존재하지 않음 " + boardUpdateRequestDto.getThumbnailId()));
-					if(!newThumb.isFlag()) {
-						File temp = new File(newThumb.getTempPath());
-						File dest = new File(newThumb.getUploadPath());
-						File origTemp = new File(newThumb.getOriginalFileTemp());
-						File origDest = new File(newThumb.getOriginalFileDest());
-						try {
-							Files.move(temp, dest);
-							Files.move(origTemp, origDest);
-						} catch (IOException e) {
-							e.printStackTrace();
+					
+					Optional<Thumbnail> newThumb = thumbnailRepository.findById(boardUpdateRequestDto.getThumbnailId());
+					if(newThumb.isPresent()) {
+						if(!newThumb.get().isFlag()) {
+							File temp = new File(newThumb.get().getTempPath());
+							File dest = new File(newThumb.get().getUploadPath());
+							File origTemp = new File(newThumb.get().getOriginalFileTemp());
+							File origDest = new File(newThumb.get().getOriginalFileDest());
+							try {
+								Files.move(temp, dest);
+								Files.move(origTemp, origDest);
+							} catch (IOException e) {
+								e.printStackTrace();
+							}
+							newThumb.get().completelySave();
 						}
-						newThumb.completelySave();
+						newThumb.get().addBoard(board);
+						thumbnailRepository.save(newThumb.get());		
 					}
-					newThumb.addBoard(board);
-					thumbnailRepository.save(newThumb);
 				}
 			}
 		}
@@ -268,14 +283,16 @@ public class BoardServiceImpl implements BoardService {
 			/*
 			 * 유저 아이디 값을 받도록 바꾸어야 한다.
 			 * */
-			
-			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId());
-			dto.setThumbnail(thumbnail.getUploadPath());
+			Optional<Thumbnail> thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId());
+			if(thumbnail.isPresent()) {
+				dto.setThumbnail(thumbnail.get().getUploadPath());
+				/***
+				 * boardsInBoardType 등 리스트를 뽑아 보낼때는 썸네일 테이블에서 리사이즈한 파일을 보내준다
+				 */ 				
+			}
 			curBoardResponseDto.add(dto);
-			/***
-			 * boardsInBoardType 등 리스트를 뽑아 보낼때는 썸네일 테이블에서 리사이즈한 파일을 보내준다
-			 */ 
 		}
+		System.out.println("=================================="+curBoardResponseDto);
 		return curBoardResponseDto;
 	}
 	
@@ -297,8 +314,8 @@ public class BoardServiceImpl implements BoardService {
 			dto.setLikesAndComments(likes, comments);
 			dto.setLiked(likedOrNot);
 			
-			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId());
-			dto.setThumbnail(thumbnail.getUploadPath());
+			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId()).orElseThrow(() -> new IllegalAccessError("썸네일이 존재하지 않음 " + board.getBoardId()));
+			dto.setThumbnail(thumbnail.getOriginalFileDest());
 			
 			curBoardResponseDto.add(dto);
 		}
@@ -323,8 +340,8 @@ public class BoardServiceImpl implements BoardService {
 			dto.setLikesAndComments(likes, comments);
 			dto.setLiked(likedOrNot);
 			
-			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId());
-			dto.setThumbnail(thumbnail.getUploadPath());
+			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId()).orElseThrow(() -> new IllegalAccessError("썸네일이 존재하지 않음 " + board.getBoardId()));
+			dto.setThumbnail(thumbnail.getOriginalFileDest());
 			
 			curBoardResponseDto.add(dto);
 		}
@@ -349,8 +366,8 @@ public class BoardServiceImpl implements BoardService {
 			dto.setLikesAndComments(likes, comments);
 			dto.setLiked(likedOrNot);
 			
-			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId());
-			dto.setThumbnail(thumbnail.getUploadPath());
+			Thumbnail thumbnail = thumbnailRepository.findByBoardBoardId(board.getBoardId()).orElseThrow(() -> new IllegalAccessError("썸네일이 존재하지 않음 " + board.getBoardId()));
+			dto.setThumbnail(thumbnail.getOriginalFileDest());
 			
 			curBoardResponseDto.add(dto);
 		}
